Changed methods to return a hash.
Refactored tests appropriately. Also got rid of two guard statements in BlueberryMuffin class.
This is because the exercise wants us to go with the more expensive option in case on 10 muffins. For example 5x2 = $33.9 BUT 8 + 2 = 34.9. We initially preferred the cheaper option, but the example we're given prefers the more expensive option.